Move some false negatives into TODO
